+++ test/filtrations/newfiltrations.jl
@@ -6,7 +6,8 @@ using SparseArrays
 using StaticArrays
 using Test
 
-include(joinpath(@__DIR__, "test-datasets.jl"))
+include("../testdatasets.jl")
+include("interfacetest.jl")
 using Ripserer: nv, edges
 
 # Julia 1.0 does not allow these to be defined inside @testset.
@@ -70,19 +71,9 @@ end
         end
     end
 
-    for m in (2, 3)
-        @testset "Results are the same for all algorithms with modulus=$m" begin
-            appa = ApparentPairsRips(projective_plane)
-            _, hom_imp = ripserer(appa; alg=:homology, implicit=true, modulus=m)
-            _, hom_exp = ripserer(appa; alg=:homology, implicit=false, modulus=m)
-            _, hom_ass = ripserer(appa; alg=:involuted, implicit=false, modulus=m)
-            _, coh_imp = ripserer(appa; alg=:cohomology, implicit=true, modulus=m, reps=1)
-            _, coh_exp = ripserer(appa; alg=:cohomology, implicit=false, modulus=m, reps=1)
-
-            @test hom_imp == hom_exp == coh_imp == coh_exp == hom_ass
-            @test representative.(hom_imp) == representative.(hom_exp)
-            @test representative.(hom_imp) == representative.(hom_ass)
-        end
+    @testset "Interface" begin
+        test_filtration(ApparentPairsRips, projective_plane; modulus=2)
+        test_filtration(ApparentPairsRips, projective_plane; modulus=3)
     end
 end
 
@@ -115,48 +106,28 @@ end
         @test d1[1].death_simplex == Simplex{2}(1, 3)
         @test simplex(only(d1[1].representative)) == Simplex{1}(3, 2)
     end
-    @testset "Results are the same for all algorithms" begin
-        appa = ApparentPairsCustom()
-        _, hom_imp = ripserer(appa; alg=:homology, implicit=true)
-        _, hom_exp = ripserer(appa; alg=:homology, implicit=false)
-        _, coh_imp = ripserer(appa; alg=:cohomology, implicit=true)
-        _, coh_exp = ripserer(appa; alg=:cohomology, implicit=false)
-
-        @test hom_imp == hom_exp == coh_imp == coh_exp
-        @test representative.(hom_imp) == representative.(hom_exp)
-    end
 end
 
 # The main idea of this test is to test postprocess_diagrams
-struct NoRepsFiltration <: Ripserer.AbstractFiltration{Int,Int} end
+struct ReversedResult <: Ripserer.AbstractFiltration{Int,Int} end
 
-function Ripserer.unsafe_simplex(::Type{Simplex{0,Int,Int}}, ::NoRepsFiltration, (v,))
+function Ripserer.unsafe_simplex(::Type{Simplex{0,Int,Int}}, ::ReversedResult, (v,))
     return Simplex{0}(v, 0)
 end
 function Ripserer.unsafe_simplex(
-    ::Type{Simplex{D,Int,Int}}, ::NoRepsFiltration, vertices
+    ::Type{Simplex{D,Int,Int}}, ::ReversedResult, vertices
 ) where {D}
     return Simplex{D}(index(vertices), 1)
 end
-Ripserer.nv(::NoRepsFiltration) = 10
-Ripserer.simplex_type(::Type{NoRepsFiltration}, D) = Simplex{D,Int,Int}
-Ripserer.edges(::NoRepsFiltration) = Simplex{1}.(10:-1:1, 1)
-function Ripserer.postprocess_diagram(::NoRepsFiltration, diagram)
-    result = PersistenceInterval[]
-    for int in diagram
-        if !hasproperty(int, :representative)
-            push!(result, PersistenceInterval(birth(int) + 1, death(int) + 1))
-        end
-    end
-    return PersistenceDiagram(sort!(result); diagram.meta...)
+Ripserer.nv(::ReversedResult) = 10
+Ripserer.simplex_type(::Type{ReversedResult}, D) = Simplex{D,Int,Int}
+Ripserer.edges(::ReversedResult) = Simplex{1}.(10:-1:1, 1)
+function Ripserer.postprocess_diagram(::ReversedResult, diagram)
+    return reverse!(diagram)
 end
 
 @testset "New filtration" begin
-    d0, d1 = ripserer(NoRepsFiltration())
-    @test d0 == [fill((1.0, 2.0), 4); fill((1.0, Inf), 6)]
-    @test d1 == []
-
-    d0, d1 = ripserer(NoRepsFiltration(); reps=true)
-    @test d0 == []
+    d0, d1 = ripserer(ReversedResult())
+    @test d0 == [fill((0.0, Inf), 6); fill((0.0, 1.0), 4)]
     @test d1 == []
 end