+++ test/filtrations/rips.jl
@@ -10,7 +10,8 @@ using Test
 using Ripserer:
     distances, births, adjacency_matrix, edges, nv, unsafe_simplex, ChainElement, Chain
 
-include("test-datasets.jl")
+include("../testdatasets.jl")
+include("interfacetest.jl")
 
 @testset "distances" begin
     for points in (
@@ -19,54 +20,6 @@ include("test-datasets.jl")
     )
         @test distances(points) ≈ [0 1 √2 1; 1 0 1 √2; √2 1 0 1; 1 √2 1 0]
         @test distances(points, Cityblock()) == [0 1 2 1; 1 0 1 2; 2 1 0 1; 1 2 1 0]
-    end
-end
-
-for sparse in (true, false)
-    @testset "With no threshold, sparse=$sparse" begin
-        filtration = Rips(Float64[0 1 2 9; 1 0 3 9; 2 3 0 4; 9 9 4 0]; sparse=sparse)
-
-        @test sprint(show, filtration) == "Rips{Int64, Float64}(nv=4, sparse=$sparse)"
-
-        @test nv(filtration) == 4
-        if issparse(adjacency_matrix(filtration))
-            @test adjacency_matrix(filtration) == [0 1 2 0; 1 0 3 0; 2 3 0 4; 0 0 4 0]
-        else
-            @test adjacency_matrix(filtration) == [0 1 2 9; 1 0 3 9; 2 3 0 4; 9 9 4 0]
-        end
-
-        @test threshold(filtration) == 4.0
-        @test births(filtration) == zeros(4)
-
-        @test adjacency_matrix(filtration) == filtration.adj
-
-        @test unsafe_simplex(filtration, Val(0), (1,)) === Simplex{0}(1, 0.0)
-        @test unsafe_simplex(filtration, Val(1), (2, 1)) === Simplex{1}(1, 1.0)
-        @test simplex(filtration, Val(2), (1, 3, 2)) === Simplex{2}(1, 3.0)
-        @test_throws ArgumentError simplex(filtration, Val(2), (1, 1, 2))
-    end
-    @testset "With threshold and index type, sparse=$sparse" begin
-        filtration = Rips{Int128}([1 2 2; 2 2 3; 2 3 2]; threshold=2, sparse=sparse)
-
-        @test sprint(show, filtration) == "Rips{Int128, Int64}(nv=3, sparse=$sparse)"
-
-        @test nv(filtration) == 3
-        @test threshold(filtration) == 2
-        if sparse
-            @test adjacency_matrix(filtration) == [1 2 2; 2 2 0; 2 0 2]
-        else
-            @test adjacency_matrix(filtration) == [1 2 2; 2 2 3; 2 3 2]
-        end
-        @test threshold(filtration) == 2
-
-        @test births(filtration) == [1, 2, 2]
-
-        @test adjacency_matrix(filtration) == filtration.adj
-
-        @test unsafe_simplex(filtration, Val(0), (1,)) === Simplex{0,Int,Int128}(1, 1)
-        @test unsafe_simplex(filtration, Val(2), (3, 2, 1)) === nothing
-        @test simplex(filtration, Val(1), (1, 2)) === Simplex{1,Int,Int128}(1, 2)
-        @test_throws ArgumentError simplex(filtration, Val(2), (0, 1, 2))
     end
 end
 
@@ -139,7 +92,7 @@ end
 end
 
 @testset "ripserer" begin
-    @testset "Rips" begin
+    @testset "Dense" begin
         @testset "Icosahedron" begin
             d0, d1, d2 = ripserer(icosahedron; dim_max=2)
             @test d0 == [fill((0.0, 1.0), 11); (0.0, Inf)]
@@ -147,9 +100,9 @@ end
             @test d2 == [(1.0, 2.0)]
         end
         @testset "Cycle with various fields" begin
-            d0_2, d1_2, d2_2, d3_2 = ripserer(Rips{Int32}(cycle); dim_max=3)
+            d0_2, d1_2, d2_2, d3_2 = ripserer(Rips{Int32}, cycle; dim_max=3)
             d0_7, d1_7, d2_7, d3_7 = ripserer(Rips(cycle); dim_max=3, modulus=7)
-            d0_r, d1_r, d2_r, d3_r = ripserer(cycle; dim_max=3, field_type=Rational{Int})
+            d0_r, d1_r, d2_r, d3_r = ripserer(cycle; dim_max=3, field=Rational{Int})
 
             @test d0_2 == d0_7 == d0_r == [fill((0, 1), size(cycle, 1) - 1); (0, Inf)]
             @test d1_2 == d1_7 == d1_r == [(1, 6)]
@@ -157,10 +110,10 @@ end
             @test d3_2 == d3_7 == d3_r == [(7, 8)]
         end
         @testset "RP2 with various fields" begin
-            _, d1_2, d2_2 = ripserer(projective_plane; dim_max=2)
-            _, d1_3, d2_3 = ripserer(projective_plane; dim_max=2, modulus=3)
-            _, d1_331, d2_331 = ripserer(projective_plane; dim_max=2, field_type=Mod{5})
-            _, d1_r, d2_r = ripserer(projective_plane; dim_max=2, field_type=Rational{Int})
+            _, d1_2, d2_2 = ripserer(Rips(projective_plane); dim_max=2)
+            _, d1_3, d2_3 = ripserer(Rips, projective_plane; dim_max=2, modulus=3)
+            _, d1_331, d2_331 = ripserer(projective_plane; dim_max=2, field=Mod{5})
+            _, d1_r, d2_r = ripserer(projective_plane; dim_max=2, field=Rational{Int})
             @test d1_2 == [(1, 2)]
             @test d2_2 == [(1, 2)]
             @test d1_3 == d1_331 == d1_r == []
@@ -179,13 +132,13 @@ end
             @test d2 == []
         end
         @testset "RP2 with various fields, threshold=1" begin
-            _, d1_2, d2_2 = ripserer(projective_plane; dim_max=2, threshold=1)
+            _, d1_2, d2_2 = ripserer(Rips, projective_plane; dim_max=2, threshold=1)
             _, d1_3, d2_3 = ripserer(projective_plane; dim_max=2, modulus=3, threshold=1)
             _, d1_331, d2_331 = ripserer(
-                projective_plane; dim_max=2, field_type=Mod{5}, threshold=1
+                projective_plane; dim_max=2, field=Mod{5}, threshold=1
             )
             _, d1_r, d2_r = ripserer(
-                projective_plane; dim_max=2, field_type=Rational{Int}, threshold=1
+                projective_plane; dim_max=2, field=Rational{Int}, threshold=1
             )
             @test d1_2 == [(1, Inf)]
             @test d2_2 == [(1, Inf)]
@@ -205,7 +158,7 @@ end
         end
     end
 
-    @testset "sparse" begin
+    @testset "Sparse" begin
         @testset "Icosahedron" begin
             d0, d1, d2 = ripserer(sparse(icosahedron); dim_max=2)
             @test d0 == [fill((0.0, 1.0), 11); (0.0, Inf)]
@@ -218,10 +171,15 @@ end
                 Rips(projective_plane; threshold=1, sparse=true); dim_max=2, modulus=3
             )
             _, d1_331, d2_331 = ripserer(
-                sparse(projective_plane); dim_max=2, field_type=Mod{5}, threshold=1
+                Rips, sparse(projective_plane); dim_max=2, field=Mod{5}, threshold=1
             )
             _, d1_r, d2_r = ripserer(
-                sparse(projective_plane); dim_max=2, field_type=Rational{Int}, threshold=1
+                Rips,
+                projective_plane;
+                sparse=true,
+                dim_max=2,
+                field=Rational{Int},
+                threshold=1,
             )
             @test d1_2 == [(1, Inf)]
             @test d2_2 == [(1, Inf)]
@@ -232,7 +190,7 @@ end
             for thresh in (nothing, 1, 0.5, 0.126)
                 data = torus_points(100)
                 r_res = ripserer(data; threshold=thresh, dim_max=2)
-                s_res_1 = ripserer(data; threshold=thresh, sparse=true, dim_max=2)
+                s_res_1 = ripserer(Rips, data; threshold=thresh, sparse=true, dim_max=2)
 
                 # Add zeros to diagonal. Adding ones first actually changes the structure of
                 # the matrix.
@@ -272,21 +230,17 @@ end
             d0, d1, d2, d3 = ripserer(cycle; dim_max=3, reps=true)
             @test d1[1].representative isa Chain
 
-            d0, d1, d2, d3 = ripserer(cycle; dim_max=3, reps=true, field_type=Rational{Int})
+            d0, d1, d2, d3 = ripserer(cycle; dim_max=3, reps=true, field=Rational{Int})
             @test d3[1].representative isa Chain
         end
         @testset "Infinite interval very low threshold" begin
-            _, d1 = ripserer(
-                cycle; dim_max=1, reps=true, threshold=1, field_type=Rational{Int}
-            )
+            _, d1 = ripserer(cycle; dim_max=1, reps=true, threshold=1, field=Rational{Int})
             rep = representative(only(d1))
             @test simplex(only(rep)) == birth_simplex(only(d1))
             @test rep isa Chain{Rational{Int},Simplex{1,Int,Int}}
         end
         @testset "Infinite interval higher threshold" begin
-            _, d1 = ripserer(
-                cycle; dim_max=1, reps=true, threshold=3, field_type=Rational{Int}
-            )
+            _, d1 = ripserer(cycle; dim_max=1, reps=true, threshold=3, field=Rational{Int})
             rep = representative(only(d1))
             @test !isempty(rep)
             @test rep isa Chain{Rational{Int},Simplex{1,Int,Int}}
@@ -305,9 +259,7 @@ end
 
     @testset "Diagram metadata" begin
         filtration = Rips(cycle)
-        d0, d1, d2, d3 = ripserer(
-            filtration; dim_max=3, reps=true, field_type=Rational{Int}
-        )
+        d0, d1, d2, d3 = ripserer(filtration; dim_max=3, reps=true, field=Rational{Int})
         @test d0.dim == 0
         @test d1.dim == 1
         @test d2.dim == 2
@@ -317,11 +269,11 @@ end
         @test d1.threshold ≡ thresh
         @test d2.threshold ≡ thresh
         @test d3.threshold ≡ thresh
-        field_type = Rational{Int}
-        @test d0.field_type ≡ field_type
-        @test d1.field_type ≡ field_type
-        @test d2.field_type ≡ field_type
-        @test d3.field_type ≡ field_type
+        field = Rational{Int}
+        @test d0.field ≡ field
+        @test d1.field ≡ field
+        @test d2.field ≡ field
+        @test d3.field ≡ field
         @test d0.filtration == filtration
         @test d1.filtration == filtration
         @test d2.filtration == filtration
@@ -366,54 +318,24 @@ end
     end
 
     @testset "Homology and explicit cohomology" begin
-        for m in (2, 17)
-            @testset "Same for projective plane with modulus=$m" begin
-                r = Rips(projective_plane)
-                _, hom_imp1, hom_imp2 = ripserer(
-                    r; alg=:homology, implicit=true, dim_max=2, modulus=m
-                )
-                _, hom_exp1, hom_exp2 = ripserer(
-                    r; alg=:homology, implicit=false, dim_max=2, modulus=m
-                )
-                _, hom_ass1, hom_ass2 = ripserer(r; alg=:involuted, dim_max=2, modulus=m)
-                _, coh_imp1, coh_imp2 = ripserer(
-                    r; implicit=true, reps=true, dim_max=2, modulus=m
-                )
-                _, coh_exp1, coh_exp2 = ripserer(
-                    r; implicit=true, reps=true, dim_max=2, modulus=m
-                )
-                @test hom_imp1 == hom_exp1 == hom_ass1 == coh_imp1 == coh_exp1
-                @test hom_imp2 == hom_exp2 == hom_ass2 == coh_imp2 == coh_exp2
-                @test representative.(hom_imp1) == representative.(hom_exp1)
-                @test representative.(hom_imp1) == representative.(hom_ass1)
-                @test representative.(hom_imp2) == representative.(hom_exp2)
-                @test representative.(hom_imp2) == representative.(hom_ass2)
-                @test representative.(coh_imp1) == representative.(coh_exp1)
-                @test representative.(coh_imp2) == representative.(coh_exp2)
-            end
-        end
         @testset "Representative cycle" begin
             res_hom = ripserer(cycle; alg=:homology, reps=true, dim_max=3)
             @test vertices.(simplex.(representative(res_hom[2][1]))) ==
                   sort!(vcat([(i + 1, i) for i in 1:17], [(18, 1)]))
         end
         @testset "Infinite intervals" begin
-            @test_broken ripserer(cycle; alg=:homology, threshold=2, implicit=true)[2][1] ==
-                         (1.0, Inf)
-            @test_broken ripserer(cycle; alg=:homology, threshold=2, implicit=false)[2][1] ==
-                         (1.0, Inf)
+            @test_broken ripserer(
+                Rips(cycle; threshold=2); alg=:homology, implicit=true
+            )[2][1] == (1.0, Inf)
+            @test_broken ripserer(
+                Rips, cycle; alg=:homology, threshold=2, implicit=false
+            )[2][1] == (1.0, Inf)
             @test ripserer(cycle; alg=:involuted, threshold=2)[2][1] == (1.0, Inf)
         end
     end
 
-    @testset "Only print to stderr and only when progress is enabled" begin
-        @suppress begin
-            @test (@capture_out ripserer(torus(16); dim_max=5)) == ""
-            @test (@capture_out ripserer(torus(16); dim_max=5, progress=true)) == ""
-
-            @test (@capture_err ripserer(torus(16); dim_max=5)) == ""
-            @test (@capture_err ripserer(torus(16); dim_max=5, progress=true)) != ""
-        end
+    @testset "Interface" begin
+        test_filtration(Rips, cycle; modulus=3, dim_max=2)
     end
 
     @testset "Errors" begin
@@ -424,9 +346,12 @@ end
             @test_throws ArgumentError ripserer(ones(5, 5); alg=:something)
         end
         @testset "Int or Float64 field type" begin
-            @test_throws ErrorException ripserer(ones(5, 5); field_type=Int)
-            @test_throws ErrorException ripserer(ones(5, 5); field_type=Float64)
-            @test_throws ErrorException ripserer(ones(5, 5); field_type=UInt8)
+            @test_throws ErrorException ripserer(ones(5, 5); field=Int)
+            @test_throws ErrorException ripserer(ones(5, 5); field=Float64)
+            @test_throws ErrorException ripserer(ones(5, 5); field=UInt8)
+        end
+        @testset "Explicit cohomology reperesentatives unsupported" begin
+            @test_throws ErrorException ripserer(cycle; implicit=false, reps=true)
         end
     end
 end